requirements.txt
flask_sqlalchemy
psycopg2-binary
flask
requests
python-dotenv
flask-login
Werkzeug